Implements Fully Auto-Answering and Bugfix (#1)
* Update README.md

* Update script.js

* Update README.md

* Update README.md

* Added auto-answer functionality

I left the original functionality intact but added the option to use ALT + S to actually answer the questions in a fully automated fashion. Code could be compressed in future if depreciated functionality was removed.

* Made alert prompts more clear

* Updates README for new features

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Fixed undefined for some questions

Certain questions were returning undefined. Turns out that was just because the mergeLists function didn't equate for commas too